Create signup middleware
