+++ app/controllers/users.js
@@ -1,15 +1,8 @@
-const emailDomain = '@wolox.com.ar',
-  bcrypt = require('bcrypt'),
+const bcrypt = require('bcrypt'),
   { databaseError, hashError, validationError } = require('../errors'),
   Users = require('../models').user,
   logger = require('../logger'),
-  saltRounds = 10,
-  minPasswordLength = 8,
-  alphanumericRegex = /^[0-9a-zA-Z]+$/;
-
-const checkValidPassword = pass => pass && pass.length >= minPasswordLength && pass.match(alphanumericRegex);
-
-const checkValidEmail = email => email && email.endsWith(emailDomain);
+  saltRounds = 10;
 
 const hashPassword = pass => bcrypt.hash(pass, saltRounds).catch(error => hashError(error.message));
 
@@ -27,16 +20,6 @@ exports.signUp = (req, res, next) => {
   const { first_name: firstName, last_name: lastName, email, password } = req.body;
   logger.info(`Creating user ${firstName}`);
 
-  if (!firstName || !lastName) {
-    return next(validationError('Missing parameters. firstName and lastName are required'));
-  }
-
-  if (!checkValidEmail(email)) {
-    return next(validationError('Invalid email'));
-  }
-  if (!checkValidPassword(password)) {
-    return next(validationError(`Invalid password. Must have ${minPasswordLength} aplhanumeric characters`));
-  }
   logger.info('Email and password correctly validated');
   return hashPassword(password)
     .then(hash => createUser(firstName, lastName, email, hash))
